+++ ClearTube/Model/VideoObject.swift
@@ -56,11 +56,11 @@ public struct VideoObject: Equatable, Decodable {
     public struct RecommendedVideoObject: Decodable {
         public var videoId: String
         public var title: String
-        public var lengthSeconds: Int32
+        public var lengthSeconds: Int
         public var videoThumbnails: [ImageObject]
         public var author: String?
         public var authorId: String?
-        public var viewCount: Int?
+        public var viewCount: Int64?
         public var viewCountText: String?
     }
